Update the example spec for the home page.
The old spec was more specific (an all zero game)
then the implementation (which summed all the pins).
Given we didn’t want to do the “simplest” thing and hard
code a return value of 0, it’s better to have a spec
that actually forces the summing of the pins.