Fix formatting of #rspec IRC channel name
On the help page the #rspec IRC channel name is listed. Using the `pre`
tag element it would show the channel name on a new line in the middle
of the sentence, using a dark background with dark text, making it
difficult to read.

Instead use the `code` tag element to show the channel name inline in
a monospaced font for it to stand out in the sentence.