Fix part of the CSS
Fix CSS generation by adding file extensions to source files.
This is needed for some files in the newer middleman versions.

Also update the paths in import statements to point to the locations
of the files using a relative path.

I disabled bourbon and neat for now. Bourbon itself is hardly used.
Neat depends on bourbon, but is so far behind it can't be easily upgraded,
as all helpers have since been replaced with a new set of helpers that are
not compatible.