Reinstall older versions of bourbon and neat
Lock bourbon and neat on the older versions of the packages and
"install" them into the project.

I've run `bourbon install --path source/stylesheets` and
`neat install --path source/stylesheets`
to install the files directly in the project. Which is why this commit
adds so many files.

While almost no bourbon mixins are used anymore, it exposes some
variables required by neat, like `$golden` and `$georgia`.

Neat is no longer maintained and I suggest it is removed from the
project and another grid system or CSS grid directly is used instead.
Then both bourbon and neat can be removed.

This solution is a quick fix so that the middleman 4 upgrade can be
completed for now. If it's replaced all the added files can be removed
again.

I replaced almost all of the other Bourbon mixins with the now
stabilized CSS properties. In earlier commits I've added a custom mixins
`clearfix`, copied from what bourbon used.