Fix build by fixing partial filtering
Fix the `ignored_sitemap_matchers` for partials.

I copied the middleman implementation of this proc and adjusted the
docs exception to work with the new method.

The `source_file` object is now a struct with multiple versions of the
source file path. The exception needed to be checked against the full
path and not the relative_path that only included the filename and not
the directories.

In the proxy file generation for the "latest" version of the docs, that
redirects to the actual latest version of the docs, I only selected the
`.html` files. This should be okay, and I tested it works. But I had to
add this filter because otherwise all the `.css` and `.js` files would
be parsed as CSS and JavaScript files. That would then break because
those included HTML and cause syntax errors during compilation.

Previously in PR #154 this solution was mentioned. I don't know the
background here. I think the site doesn't also need to redirect the
`.css` and `.js` files. Only the `.html` files redirection should be
enough, as it will load the assets from the actual versioned docs.